+++ samples/qa-virtualgridlistnative-preserve-focus/src/index.js
@@ -1,11 +1,10 @@
 import {Provider} from 'react-redux';
-import {render} from 'react-dom';
+import {createRoot} from 'react-dom/client';
 
 import App from './App';
 import configureStore from './store';
 
 const store = configureStore();
-
 const appElement = (
 	<Provider store={store}>
 		<App />
@@ -14,7 +13,10 @@ const appElement = (
 
 // In a browser environment, render the app to the document.
 if (typeof window !== 'undefined') {
-	render(appElement, document.getElementById('root'));
+	const container = document.getElementById('root');
+	const root = createRoot(container);
+
+	root.render(appElement);
 }
 
 export default appElement;