+++ CHANGELOG.md
@@ -2,11 +2,28 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
-## [unreleased]
+## [2.5.0-alpha.1] - 2022-04-15
+
+- Update dependencies including React 18.0.0
+
+### Changed
+
+- `sandstone/DatePicker` and `sandstone/TimePicker` to not show press effect on touch input
+
+### Fixed
+
+- `sandstone/Scroller` and `sandstone/VirtualList` to focus the topmost element after scroll in pointer mode
+
+## [2.1.4] - 2022-03-24
 
 ### Added
 
 - `sandstone/Icon` public class name `icon`
+- `sandstone/Scroller` and `sandstone/VirtualList` prop `data-webos-voice-focused`, `data-webos-voice-disabled`, and `data-webos-voice-group-label`
+
+### Fixed
+
+- `sandstone/WizardPanels` to provide a way to prevent focusing on Panel again by allowing preventDefault when `onTransition` and `onWillTransition`
 
 ## [2.1.3] - 2022-03-07